pyproject.toml
[project]
name = "doctotxt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "einops>=0.8.2",
    "pdf2image>=1.17.0",
    "python-dotenv>=1.2.1",
    "timm>=1.0.24",
    "torch>=2.10.0",
    "transformers==4.49.0",
]
